AD-68: Alle Tests schlagen fehl, Problem lag in Repository interface findOne war falsch.

+++ src/main/java/de/hs/furtwangen/bam/spots/service/UserService.java
@@ -8,7 +8,7 @@ public interface UserService {
 	
 	public User findByUsername(String username);
 	
-	public User findOne(int id);
+	public User findOne(long id);
 	
 	public void save(User user);
